pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tunnelbar"
version = "0.1.0"
description = "macOS menubar SSH tunnel manager"
requires-python = ">=3.11"
dependencies = [
    "rumps>=0.4.0",
    "pyyaml>=6.0",
]

[project.scripts]
tunnelbar = "tunnelbar.app:main"

[project.optional-dependencies]
dev = ["py2app>=0.28.9"]

[tool.setuptools.package-data]
tunnelbar = ["resources/*.png", "resources/*.icns"]
